new code edit
control/cpanel/pages/main/cards.php

mt_rand() couldn't run because the function parameters were a negetive integers thefore I increased the length of the numbers

changed to __autoload() to spl_autoload_register() because __autoload() does not work

Check if email is already registered before inserting

because of "PHP Error: Cannot use object of type stdClass as array (array and object issues)" I had to change from [""] to -> because the code is using data in the form of an array holding objects. Objects and arrays are not the same, and because of this they use different syntaxes for accessing data.

Errors:
1. After registering for adminsion page redirect to a non available url - http://127.0.0.1/hisp.portal.com/prospectStudent/dashboard/profile/editprofile?complete
2.